+++ FirstApp HomeWork/Base.lproj/Main.storyboard
@@ -204,10 +204,10 @@
             </objects>
             <point key="canvasLocation" x="2615.9420289855075" y="-267.1875"/>
         </scene>
-        <!--все фото-->
+        <!--Friends Foto Collection View Controller-->
         <scene sceneID="0a8-IT-Fyc">
             <objects>
-                <collectionViewController id="EAO-Oy-bMy" customClass="FriendsFotoCollectionViewController" customModule="FirstApp_HomeWork" customModuleProvider="target" sceneMemberID="viewController">
+                <collectionViewController id="EAO-Oy-bMy" userLabel="Friends Foto Collection View Controller" customClass="FriendsFotoCollectionViewController" customModule="FirstApp_HomeWork" customModuleProvider="target" sceneMemberID="viewController">
                     <collectionView key="view" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleAspectFit" dataMode="prototypes" id="5AR-8E-CWT">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
