+++ FirstApp HomeWork/Controllers/FriendsTableViewController.swift
@@ -10,10 +10,18 @@ import UIKit
 class FriendsTableViewController: UITableViewController {
     
     
-    @IBOutlet var friendsTableView: UITableView!
+    @IBOutlet var friendsTableView: UITableView! {
+    
+    didSet {
+        friendsTableView.delegate = self
+        friendsTableView.dataSource = self
+    }
+    }
+    
     
     var friendsArray = [User]()
     var savedObject: Any?
+    let dataSetingsUser = DataSettings()
     
     let reuseIdentifierUserTableCell = "UserCell"
     let segueIdentifierToFotoController = "segueIdentifierToFotoController"
@@ -49,13 +57,15 @@ class FriendsTableViewController: UITableViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        friendsArray = dataSetingsUser.setupUser()
+        
         self.clearsSelectionOnViewWillAppear = false
         self.navigationItem.rightBarButtonItem = self.editButtonItem
-        self.tableView.dataSource = self
-        self.tableView.delegate = self
+//        self.tableView.dataSource = self
+//        self.tableView.delegate = self
         
         
-        self.tableView.register(UINib(nibName: "FriendsTableViewCell", bundle: nil), forCellReuseIdentifier: reuseIdentifierUserTableCell)
+        friendsTableView.register(UINib(nibName: "FriendsTableViewCell", bundle: nil), forCellReuseIdentifier: reuseIdentifierUserTableCell)
     }
 
     // MARK: - Table view data source.
@@ -82,18 +92,18 @@ class FriendsTableViewController: UITableViewController {
         return true
     }
     
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        guard let cell = tableView.cellForRow(at: indexPath) as? FriendsTableViewCell,
-              let cellObject = cell.savedObject as? User else { return }
-        performSegue(withIdentifier: segueIdentifierToFotoController, sender: cellObject)
-    }
-    
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         friendsArray.remove(at: indexPath.row)
         tableView.deleteRows(at: [indexPath], with: .automatic)
     }
     
     // MARK: - Table view Delegate
+    
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        guard let cell = tableView.cellForRow(at: indexPath) as? FriendsTableViewCell,
+              let cellObject = cell.savedObject as? User else { return }
+        performSegue(withIdentifier: segueIdentifierToFotoController, sender: cellObject)
+    }
     
     override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         return 100
