+++ FirstApp HomeWork/Controllers/NewsViewController.swift
@@ -20,28 +20,30 @@ class NewsViewController: UIViewController {
     
     private var newsArray = [News]()
     
-    private func setupNews() -> [News] {
-        
-        var resultArray = [News]()
-        
-        let newsOne = News(newsText: "По итогам первого дня официальных тестов MotoGP на трассе Сепанг итоговый протокол возглавили два пилота Aprilia. Частично это не есть неожиданность, поскольку благодаря концессиям у обоих пилотов была возможность выехать на малазийский трек раньше остальных. Однако результат лидера Алейша Эспаргаро всё равно впечатляет, ведь он проехал лучший круг за 1’58.371 и уступил всего одну десятую рекорду трассы. И это не его максимум.", newsImage: UIImage(named: "n_1"))
-        
-        resultArray.append(newsOne)
-        
-        let newsTwo = News(newsText: "Команда Gresini Racing официально сообщила о печальном известии. К сожалению, чемпион мира 125сс (1985, 1987) и руководитель команды MotoGP Фаусто Грезини (основана в 1997-м) скончался после почти двухмесячной госпитализации из-за осложнений, вызванных коронавирусом.", newsImage: UIImage(named: "n_2"))
-        
-        resultArray.append(newsTwo)
-        
-        let newsThree = News(newsText: "12 января Маверику Виньялесу, одной из самых ярких персон современного MotoGP, исполнилось 27 лет. Несмотря на столь юный возраст, ТопГан успел повыступать в цветах трёх разных производителей, и в сезоне-2022 он продолжит выступать за Aprilia. Маверика всегда считали одним из самых одарённых, однако нестабильные результаты ещё не позволили Маку побороться за престижную корону.", newsImage: UIImage(named: "n_3"))
-        
-        resultArray.append(newsThree)
-        
-        let newsFour = News(newsText: "Тесты в Сепанге — возможность для ощущения сладостного предвкушения предстоящего сезона MotoGP. Хотя для гонщиков и производителей тесты в основном является возможностью для тонкой настройки мотоцикла, на котором они проведут сезон-2022. Пилот Suzuki Алекс Ринс доволен тем, как провёл первый день малазийских тестов, заняв третье место, хотя, как обычно, ещё есть куда расти.", newsImage: UIImage(named: "n_4"))
-        
-        resultArray.append(newsFour)
-        
-        return resultArray
-    }
+    private let dataSetingsNews = DataSettings()
+    
+//    private func setupNews() -> [News] {
+//        
+//        var resultArray = [News]()
+//        
+//        let newsOne = News(newsText: "По итогам первого дня официальных тестов MotoGP на трассе Сепанг итоговый протокол возглавили два пилота Aprilia. Частично это не есть неожиданность, поскольку благодаря концессиям у обоих пилотов была возможность выехать на малазийский трек раньше остальных. Однако результат лидера Алейша Эспаргаро всё равно впечатляет, ведь он проехал лучший круг за 1’58.371 и уступил всего одну десятую рекорду трассы. И это не его максимум.", newsImage: UIImage(named: "n_1"))
+//        
+//        resultArray.append(newsOne)
+//        
+//        let newsTwo = News(newsText: "Команда Gresini Racing официально сообщила о печальном известии. К сожалению, чемпион мира 125сс (1985, 1987) и руководитель команды MotoGP Фаусто Грезини (основана в 1997-м) скончался после почти двухмесячной госпитализации из-за осложнений, вызванных коронавирусом.", newsImage: UIImage(named: "n_2"))
+//        
+//        resultArray.append(newsTwo)
+//        
+//        let newsThree = News(newsText: "12 января Маверику Виньялесу, одной из самых ярких персон современного MotoGP, исполнилось 27 лет. Несмотря на столь юный возраст, ТопГан успел повыступать в цветах трёх разных производителей, и в сезоне-2022 он продолжит выступать за Aprilia. Маверика всегда считали одним из самых одарённых, однако нестабильные результаты ещё не позволили Маку побороться за престижную корону.", newsImage: UIImage(named: "n_3"))
+//        
+//        resultArray.append(newsThree)
+//        
+//        let newsFour = News(newsText: "Тесты в Сепанге — возможность для ощущения сладостного предвкушения предстоящего сезона MotoGP. Хотя для гонщиков и производителей тесты в основном является возможностью для тонкой настройки мотоцикла, на котором они проведут сезон-2022. Пилот Suzuki Алекс Ринс доволен тем, как провёл первый день малазийских тестов, заняв третье место, хотя, как обычно, ещё есть куда расти.", newsImage: UIImage(named: "n_4"))
+//        
+//        resultArray.append(newsFour)
+//        
+//        return resultArray
+//    }
     
     
     
@@ -54,7 +56,7 @@ class NewsViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        newsArray = setupNews()
+        newsArray = dataSetingsNews.setupNews()
         
         collectionView.backgroundColor = #colorLiteral(red: 0.8039215803, green: 0.8039215803, blue: 0.8039215803, alpha: 1)
