+++ FirstApp HomeWork/Model/Structure/Group.swift
