+++ FirstApp HomeWork/Model/Structure/User.swift
