--- FirstApp HomeWork/View/TestButton.swift
@@ -1,20 +1,0 @@
-//
-//  TestButton.swift
-//  FirstApp HomeWork
-//
-//  Created by Евгений Доброволец on 06.01.2022.
-//
-
-import UIKit
-
-class TestButton: UIView {
-
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
-    }
-    */
-
-}
